adding second screen and the navigation to the project

+++ app/src/main/java/com/example/specialmovies/data/remote/retrofit/WebServices.kt
@@ -16,7 +16,8 @@ interface WebServices {
 
     @GET("movie/{id}")
     suspend fun getMovieDetails(
-        @Path("id") movieId: Int,
+        @Path("id") movieId: Long,
         @Query("api_key") apiKey: String
     ): Response<MovieDetailsResponse>
+
 }
